Avoid cache file read/write race condition
There were frequent log error entries reporting that the cache files
were invalid JPEG files.  This probably happened while trying to read
a cache JPEG file while one of the download threads was writing the
file.

The effect of this error is that the tile data cannot be obtained and
the default (grey) texture gets returned to X-Plane, resulting in
parts of the scenery rendered with the plain grey texture and then
"popping" to the correct texture when getting closer.

Temporarily rename the JPEG file while it is being written to use a
cache file name that AO will not look for, to prevent reading it before
it is ready.