Wait for all Tile chunks instead of just one
get_img() submits requests to download all the chunks for a Tile, but
only waited for the first one to complete.  This could lead to using
unloaded chunks to populate the Tile which show up as grey textures
in X-Plane.

The code here was changed/commented out to only wait on the first
chunk, probably in an effor to reduce stuttering.  Waiting on each
chunk could possibly result in taking longer to return a Tile to
X-Plane and cause stuttering.

This commit reverts the code back to waiting on all the chunks.  This
greatly reduces the chances of grey textures, at the possible expense
of stuttering.